move reports slice to activity section
