+++ apps/fishing-map/features/user/UserReports.tsx
@@ -15,7 +15,7 @@ import {
 } from 'features/reports/areas/reports.slice'
 import { REPORT } from 'routes/routes'
 import { selectUserReports } from 'features/user/selectors/user.permissions.selectors'
-import { resetReportData } from 'features/reports/areas/report.slice'
+import { resetReportData } from 'features/reports/activity/reports-activity.slice'
 import { resetWorkspaceSlice } from 'features/workspace/workspace.slice'
 import styles from './User.module.css'
