+++ apps/fishing-map/reducers.ts
@@ -19,7 +19,7 @@ import mapControlsReducer from 'features/map/controls/map-controls.slice'
 import mapReducer from 'features/map/map.slice'
 import modalsReducer from 'features/modals/modals.slice'
 import regionsReducer from 'features/regions/regions.slice'
-import reportReducer from 'features/reports/areas/report.slice'
+import reportReducer from 'features/reports/activity/reports-activity.slice'
 import reportsReducer from 'features/reports/areas/reports.slice'
 import resourcesReducer from 'features/resources/resources.slice'
 import searchReducer from 'features/search/search.slice'
